[BUGFIX] Allow image to zoom under navigation bar
In the previous version, the size of the scroll view does not cover
the area under the navigation bar. When user hides the navigation bar,
the zoomed image cannot cover the navigation bar area and there is
always a black area on the top.

The calculation of the scroll view frame is updated so that the scroll
view frame can cover the area under the navigation bar.

Resolves:
Releases: 0.0.2